feat(agents): add ACP adapter to agents
